disable obstacle drawing in debug mode
