+++ packages/blocks/src/lib/navigation-map/components/draw-tool.svelte
