+++ packages/blocks/src/lib/navigation-map/components/map.svelte
@@ -2,7 +2,7 @@
 import { Button, Icon, Radio, Tooltip } from '@viamrobotics/prime-core';
 import type { Map } from 'maplibre-gl';
 import { MapLibre } from '$lib';
-import { view } from '../stores';
+import { environment, view } from '../stores';
 import ObstacleLayer from './obstacle-layer.svelte';
 import RobotMarker from './robot-marker.svelte';
 import CenterInputs from './center-inputs.svelte';
@@ -10,7 +10,7 @@ import Nav from './nav/index.svelte';
 import Waypoints from './waypoints.svelte';
 import ObstaclesLegend from './nav/obstacles-legend.svelte';
 
-/** The Geopose of a robot base. */
+/** The Geo-pose of a robot base. */
 export let baseGeoPose: { lng: number; lat: number } | undefined = undefined;
 
 const minPitch = 0;
@@ -59,20 +59,22 @@ const toggleTileset = () => {
     />
 
     <div class="absolute right-12 top-2.5 z-10 flex items-center gap-2">
-      <Tooltip>
-        <div class="rounded-full bg-white">
-          <Icon
-            size="lg"
-            name="help-circle-outline"
-          />
-        </div>
-        <div
-          class="flex flex-col p-2"
-          slot="description"
-        >
-          <ObstaclesLegend />
-        </div>
-      </Tooltip>
+      {#if $environment === 'configure'}
+        <Tooltip>
+          <div class="rounded-full bg-white">
+            <Icon
+              size="lg"
+              name="help-circle-outline"
+            />
+          </div>
+          <div
+            class="flex flex-col p-2"
+            slot="description"
+          >
+            <ObstaclesLegend />
+          </div>
+        </Tooltip>
+      {/if}
 
       <Button on:click={toggleTileset}>
         {satellite ? 'Map' : 'Satellite'}
